Adapt code to Maven and Java 8

+++ src/main/scala/ExponentialBackOff.scala
@@ -8,7 +8,7 @@ class ExponentialBackOff(maxAttempts: Int = 3,
   extends RetryPolicy(maxAttempts) {
 
   override def run(): Unit = {
-    val wait = baseTimer.toMillis * Math.pow(2, count - 1).toLong + Random.nextLong(jitter.toMillis)
+    val wait = baseTimer.toMillis * Math.pow(2, count - 1).toLong + Random.nextInt(jitter.toMillis.toInt)
     super.run()
     logger.warn(s"Waiting for ${wait}ms")
     Thread.sleep(wait)
